Reword: Calculator: Support perfect square root
This patch allows the calculator to compute square root at any level of
precision required. It also includes a new way to store results. Before,
only the last number was stored, nonetheless this simple solution can't
be used to store irrational number as they might need to be
re-calculated with a greater precision. Now, we store all in-between
calculation to redo then with a smaller error if needed.