LibThreading: Don't assume that the `EventLoop` is still alive
While executing a `BackgroundAction`, we used a reference to the
original `EventLoop`. However, we didn't ensure that the object was
still alive. This situation led us to UAF when some conditions are met.

This patch makes `EventLoop` weakable, and makes `BackgroundAction` use
a `WeakPtr` in order to check if the `EventLoop` is still alive.

Keep in mind that this solution is still sensitive to a TOCTOU window,
meaning that the UAF still possible. But making `EventLoop` thread-safe
while keeping a way to add jobs in main thread's `EventLoop` is way more
complicated than that.