LibCore+Userland+Ladybird: Explicitly manage Core::Process's end-state
This removes the implicit `disown`ing of the Process in the destructor
in favor of up front requests in the constructor. We also force users
to exclusively and explicitly disown, wait or take the pid of the child.
I hope that it will a good way to prevent API misuses, like calling C
functions with the pid or `wait_for_termination` on a disowned child.

It unfortunately prevents the usage of the "old" API with
KeepAsChild::Yes, this is why there are some refactorings that come
with the patch.

This also should make it possible to implement KeepAsChild::No on other
platforms.